Move PageStorage::Config up
