+++ dbms/src/Storages/Page/V2/PageStorage.h
@@ -87,10 +87,10 @@ public:
 public:
     PageStorage(String name,
                 PSDiskDelegatorPtr delegator, //
-                const Config & config_,
+                const PageStorageConfig & config_,
                 const FileProviderPtr & file_provider_,
                 bool no_more_insert_ = false);
-    ~PageStorage() = default;
+    ~PageStorage() override = default;
 
     void restore() override;
 
@@ -200,6 +200,7 @@ public:
 
 #ifndef NDEBUG
     // Just for tests, refactor them out later
+    // clang-format off
     DB::PageStorage::SnapshotPtr getSnapshot() { return getSnapshot(""); }
     void write(DB::WriteBatch && wb) { return writeImpl(std::move(wb), nullptr); }
     DB::PageEntry getEntry(PageId page_id) { return getEntryImpl(TEST_NAMESPACE_ID, page_id, nullptr); }
@@ -212,6 +213,7 @@ public:
     PageMap read(const std::vector<PageReadFields> & page_fields) { return readImpl(TEST_NAMESPACE_ID, page_fields, nullptr, nullptr, true); }
     void traverse(const std::function<void(const DB::Page & page)> & acceptor) { return traverseImpl(acceptor, nullptr); }
     bool gc() { return gcImpl(false, nullptr, nullptr); }
+    // clang-format on
 #endif
 
 #ifndef DBMS_PUBLIC_GTEST
