+++ dbms/src/Storages/Page/V2/gc/LegacyCompactor.h
@@ -68,7 +68,7 @@ private:
     PSDiskDelegatorPtr delegator;
     FileProviderPtr file_provider;
 
-    const PageStorage::Config & config;
+    const PageStorageConfig & config;
 
     Poco::Logger * log;
     Poco::Logger * page_file_log;
