+++ dbms/src/Storages/Page/V2/tests/gtest_page_storage_multi_paths.cpp
@@ -79,7 +79,7 @@ protected:
     }
 
 protected:
-    PageStorage::Config config;
+    PageStorageConfig config;
     std::shared_ptr<PageStorage> storage;
     const FileProviderPtr file_provider;
 };
