+++ dbms/src/Storages/Page/V3/PageStorageImpl.cpp
@@ -39,7 +39,7 @@ namespace PS::V3
 PageStorageImpl::PageStorageImpl(
     String name,
     PSDiskDelegatorPtr delegator_,
-    const Config & config_,
+    const PageStorageConfig & config_,
     const FileProviderPtr & file_provider_)
     : DB::PageStorage(name, delegator_, config_, file_provider_)
     , log(Logger::get("PageStorage", name))
