+++ dbms/src/Storages/Page/V3/tests/gtest_page_storage.cpp
@@ -63,7 +63,7 @@ public:
         page_storage->restore();
     }
 
-    std::shared_ptr<PageStorageImpl> reopenWithConfig(const PageStorage::Config & config_)
+    std::shared_ptr<PageStorageImpl> reopenWithConfig(const PageStorageConfig & config_)
     {
         auto path = getTemporaryPath();
         delegator = std::make_shared<DB::tests::MockDiskDelegatorSingle>(path);
@@ -77,7 +77,7 @@ protected:
     FileProviderPtr file_provider;
     std::unique_ptr<StoragePathPool> path_pool;
     PSDiskDelegatorPtr delegator;
-    PageStorage::Config config;
+    PageStorageConfig config;
     std::shared_ptr<PageStorageImpl> page_storage;
 
     std::list<PageDirectorySnapshotPtr> snapshots_holder;
@@ -788,7 +788,7 @@ try
     }
 
     // Restore, the broken meta should be ignored
-    page_storage = reopenWithConfig(PageStorage::Config{});
+    page_storage = reopenWithConfig(PageStorageConfig{});
 
     {
         size_t num_pages = 0;
@@ -817,7 +817,7 @@ try
     }
 
     // Restore again, we should be able to read page 1
-    page_storage = reopenWithConfig(PageStorage::Config{});
+    page_storage = reopenWithConfig(PageStorageConfig{});
 
     {
         size_t num_pages = 0;
@@ -892,7 +892,7 @@ try
     }
 
     // Restore again, we should be able to read page 1
-    page_storage = reopenWithConfig(PageStorage::Config{});
+    page_storage = reopenWithConfig(PageStorageConfig{});
 
     {
         size_t num_pages = 0;
@@ -1565,7 +1565,7 @@ TEST_F(PageStorageTest, EntryTagAfterFullGC)
 try
 {
     {
-        PageStorage::Config config;
+        PageStorageConfig config;
         config.blob_heavy_gc_valid_rate = 1.0; /// always run full gc
         page_storage = reopenWithConfig(config);
     }
@@ -1615,7 +1615,7 @@ TEST_F(PageStorageTest, DumpPageStorageSnapshot)
 try
 {
     {
-        PageStorage::Config config;
+        PageStorageConfig config;
         config.blob_heavy_gc_valid_rate = 1.0; /// always run full gc
         config.wal_roll_size = 1 * 1024 * 1024; /// make the wal file more easy to roll
         config.wal_max_persisted_log_files = 10; /// avoid checkpoint when gc
@@ -1669,7 +1669,7 @@ try
     ASSERT_TRUE(done_snapshot);
 
     {
-        PageStorage::Config config;
+        PageStorageConfig config;
         page_storage = reopenWithConfig(config);
     }
 
@@ -1681,7 +1681,7 @@ TEST_F(PageStorageTest, DumpPageStorageSnapshotWithRefPage)
 try
 {
     {
-        PageStorage::Config config;
+        PageStorageConfig config;
         config.blob_heavy_gc_valid_rate = 1.0; /// always run full gc
         config.wal_roll_size = 1 * 1024 * 1024; /// make the wal file more easy to roll
         config.wal_max_persisted_log_files = 10; /// avoid checkpoint when gc
@@ -1745,7 +1745,7 @@ try
     ASSERT_TRUE(done_snapshot);
 
     {
-        PageStorage::Config config;
+        PageStorageConfig config;
         page_storage = reopenWithConfig(config);
     }
