+++ dbms/src/Storages/Page/workload/HighValidBigFileGC.h
@@ -58,7 +58,7 @@ public:
         {
             stop_watch.start();
 
-            DB::PageStorage::Config config;
+            DB::PageStorageConfig config;
             config.file_max_size = 8ULL * DB::GB;
             config.file_roll_size = 8ULL * DB::GB;
             initPageStorage(config, name());
@@ -79,7 +79,7 @@ public:
 
         // Generate normal data in the same Pagefile
         {
-            DB::PageStorage::Config config;
+            DB::PageStorageConfig config;
             config.file_max_size = DB::PAGE_FILE_MAX_SIZE;
             config.file_roll_size = DB::PAGE_FILE_ROLL_SIZE;
             initPageStorage(config, name());
