+++ dbms/src/Storages/Page/workload/ThousandsOfOffset.h
@@ -70,7 +70,7 @@ private:
     void run() override
     {
         pool.addCapacity(1 + options.num_writers + options.num_readers);
-        DB::PageStorage::Config config;
+        DB::PageStorageConfig config;
         initPageStorage(config, name());
 
         metrics_dumper = std::make_shared<PSMetricsDumper>(1);
